Data/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
scikit-learn==1.3.2
joblib==1.4.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.1
tmdbv3api==1.9.0
deep-translator==1.11.4
aiohttp==3.9.3